requirements_backtesting.txt
# Core dependencies
nba-api==1.1.14
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=0.21.0

# OddsAPI and betting-specific
oddsapi-python>=0.1.7  # Official OddsAPI client
sportsbooklib>=1.2.1   # For odds conversion

# Data processing
ratelimit>=2.2.1       # For API rate limiting
diskcache>=5.6.0       # For caching responses
tqdm>=4.65.0           # Progress bars

# Analysis and backtesting
scipy>=1.10.0          # For correlation calculations
scikit-learn>=1.2.0    # For advanced metrics
statsmodels>=0.13.0    # For statistical testing

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.13.0

# Development tools (optional)
pytest>=7.2.0
black>=23.3.0
mypy>=1.0.0